+++ packages/prop-house-webapp/src/components/ProposalModal/index.tsx
@@ -25,7 +25,7 @@ import {
 import { dispatchSortProposals, SortType } from '../../utils/sortingProposals';
 import LoadingIndicator from '../LoadingIndicator';
 import NotFound from '../NotFound';
-import ProposalHeaderAndBody from '../ProposalHeaderAndBody/ProposalHeaderAndBody';
+import ProposalHeaderAndBody from '../ProposalHeaderAndBody';
 import ProposalModalFooter from '../ProposalModalFooter';
 import ErrorModal from '../ErrorModal';
 import VoteConfirmationModal from '../VoteConfirmationModal';
